Consume in stead of transform
Streams.mapM_ would apply the action on the values in the InputStream,
but only when the values are consumed by another function. As we used
it as final (lazy) transformation, before discarding the stream, nothing
really happened (due to laziness). The map therefore immediatly exited
and closed the socket.

By replacing the mapM_ with an OutputStream, values are actually
processed as long as they are supplied.